third_party/WebKit/LayoutTests/platform/win/external/wpt/IndexedDB/idbindex_openKeyCursor2-expected.txt

This is a testharness.js-based test.
PASS IDBIndex.openKeyCursor() - throw InvalidStateError when the index is deleted 
Harness: the test ran to completion.
